+++ src/main/java/kr/sesacjava/swimtutor/leveltest/dto/LevelDTO.java
@@ -1,0 +1,15 @@
+package kr.sesacjava.swimtutor.leveltest.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class LevelDTO {
+    private String style;
+    private String level;
+}
